+++ include/Game/Camera/CameraParamChunkID.hpp
@@ -10,14 +10,14 @@ public:
     bool operator>(const CameraParamChunkID &) const;
     bool operator==(const CameraParamChunkID &) const;
 
-    virtual char *getBuffer(unsigned long);
+    virtual char *getBuffer(u32);
 
-    bool equals(long, const char *) const;
-    void createCubeID(long, unsigned short);
-    void createGroupID(long, const char *, unsigned long, unsigned long);
-    void createOtherID(long, const char *);
-    void createEventID(long, const char *);
-    void createStartID(long, unsigned short);
+    bool equals(s32, const char *) const;
+    void createCubeID(s32, u16);
+    void createGroupID(s32, const char *, u32, u32);
+    void createOtherID(s32, const char *);
+    void createEventID(s32, const char *);
+    void createStartID(s32, u16);
 
     s8 mZoneID;     // 0x4
     u8 _5[3];
