+++ include/Game/Camera/CameraPolygonCodeUtil.hpp
@@ -1,6 +1,8 @@
 #pragma once
 
+#include <revolution.h>
+
 namespace MR {
-    void initCameraCodeCollection(const char *, long);
+    void initCameraCodeCollection(const char *, s32);
     void termCameraCodeCollection();
 }
