+++ include/Game/Camera/DotCamParams.hpp
@@ -15,8 +15,8 @@ public:
     virtual u32 getVersion() const = 0;
     virtual bool hasMoreChunk() const = 0;
     virtual void nextToChunk() = 0;
-    virtual bool getValueInt(const char *, long *) = 0;
-    virtual bool getValueFloat(const char *, float *) = 0;
+    virtual bool getValueInt(const char *, s32 *) = 0;
+    virtual bool getValueFloat(const char *, f32 *) = 0;
     virtual bool getValueVec(const char *, TVec3f *) = 0;
     virtual bool getValueString(const char *, const char **) = 0;
 };
@@ -29,8 +29,8 @@ public:
     virtual u32 getVersion() const;
     virtual bool hasMoreChunk() const;
     virtual void nextToChunk();
-    virtual bool getValueInt(const char *, long *);
-    virtual bool getValueFloat(const char *, float *);
+    virtual bool getValueInt(const char *, s32 *);
+    virtual bool getValueFloat(const char *, f32 *);
     virtual bool getValueVec(const char *, TVec3f *);
     virtual bool getValueString(const char *, const char **);
